test(routing): add new in-depth unit tests for general message routing (#669)
Adds unit tests (both example-based and property-based) that tests
in-depth the general message routing functionality. Because of these
tests, some edge-case issues were found and fixed here as well. Mostly
related to matching functionality and correct error summary logging.

> [!NOTE]  
> The example-based unit tests are for human consumption and are mostly
there to showcase in the most simple manner how the message routing
functionality works. The property-based unit tests are there to in-depth
generate 100 tests / property so that we can randomize scenarios and be
more certain of stable functionality. Combining them is key in test
communication.

These tests will act as a substitute for the current ones that verify
only deprecated functionalty.