fix(routing): post-skipped handler could alter final error (#677)
When a message handler failed to process the message, the next handler
will be checked, but if that handler skipped/unmatched against the
received message, the final error will be altered as well.

This results in unwanted error-handling. This PR updates this subtle
change by tracking whether a message went through a failed handler, even
though, an unrelated handler was registered the failed one.

Closes #676